extract global types
this way they don’t get packaged
