+++ src/types.ts
@@ -24,9 +24,9 @@ export type GuardianAnalyticsConfig = {
 	trackers: Record<string, string>;
 };
 
-export type GuardianWindowConfig = {
+export interface GuardianWindowConfig {
 	googleAnalytics?: GuardianAnalyticsConfig;
-};
+}
 
 export type GoogleTagParams = unknown;
 export type GoogleTrackConversionObject = {
@@ -35,26 +35,8 @@ export type GoogleTrackConversionObject = {
 	google_remarketing_only: boolean;
 };
 
-declare global {
-	interface Navigator {
-		connection?: {
-			downlink: number;
-			effectiveType: string;
-		};
-	}
-	interface Window {
-		google_trackConversion?: (arg0: GoogleTrackConversionObject) => void;
-		google_tag_params?: GoogleTagParams;
-		_brandmetrics?: Array<{
-			cmd: string;
-			val: Record<string, unknown>;
-		}>;
-		googletag?: googletag.Googletag;
-		guardian: {
-			commercialTimer?: EventTimer;
-			config?: GuardianWindowConfig;
-		};
-		ga: UniversalAnalytics.ga;
-		readonly navigator: Navigator;
-	}
+export interface GuardianCommercial {
+	commercialTimer?: EventTimer;
+	config?: GuardianWindowConfig;
+	[x: string]: unknown;
 }
